Test refactored logic on ensembl_protists
